Refactor: 모달 type 분리 및 css 수정
