+++ styles/agenda/modal/AgendaModalProvider.module.scss
@@ -1,0 +1,9 @@
+@import 'styles/agenda/common.scss';
+
+.backdrop {
+  @include backdrop;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
